Switch to a better privileges check
The capability edit_plugins was originally intended for when editing the files of a plugin (PHP, JS) though the online file editor. A problem with this is if you have for security reasons added define( ‘DISALLOW_FILE_EDIT’, true ); to your website no one has the edit_plugins rights. As described here: https://wordpress.org/support/topic/improved-way-to-check-for-sufficient-priviliges/